+++ src/skyrimsavegame.h
@@ -19,8 +19,8 @@ public:
 
 protected:
   // Fetch easy-to-access information.
-  void fetchInformationFields(FileWrapper& wrapper, unsigned long& saveNumber,
-                              QString& playerName, unsigned short& playerLevel,
+  void fetchInformationFields(FileWrapper& wrapper, uint32_t& saveNumber,
+                              QString& playerName, uint16_t& playerLevel,
                               QString& playerLocation, FILETIME& creationTime) const;
 
   std::unique_ptr<DataFields> fetchDataFields() const override;
